fix: 동일 시간 선택시 api 요청하지 않도록, toast띄우지 않도록
